bet-maker/pyproject.toml
[tool.poetry]
name = "bet-maker"
version = "0.1.0"
description = "The bet-maker service is designed for betting on events by users."
authors = ["Ilya Viarbilo <verbiloilyaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.2"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
pydantic-settings = "^2.6.0"
aerich = "^0.7.2"
aio-pika = "^9.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
